combined the component and reducer for the units view table for both column and section
