+++ dacite/pages/column/[col_id]/index.ts
@@ -1,22 +1,15 @@
-import React, { useReducer } from "react";
 import h from "@macrostrat/hyper";
 import { GetServerSideProps } from "next";
 import pg, {
   BasePage,
   EditButton,
   createUnitBySections,
-  MinEditorToggle,
   UnitsView,
   ColSectionI,
-  ColSectionsTable,
-  ColSecUnitsTable,
-  ColumnPageBtnMenu,
-  UnitEditorModel,
   getIdHierarchy,
   QueryI,
+  UnitSectionTable,
 } from "~/index";
-import { columnReducer } from "~/components/column/reducer";
-import { DropResult } from "react-beautiful-dnd";
 
 export const getServerSideProps: GetServerSideProps = async (ctx) => {
   const {
@@ -64,66 +57,7 @@ export default function Columns(props: {
   query: QueryI;
   sections: { [section_id: number | string]: UnitsView[] }[];
 }) {
-  const { col_id, colSections, column, query } = props;
-
-  const [state, dispatch] = useReducer(columnReducer, {
-    sections: props.sections,
-    mergeIds: [],
-    divideIds: [],
-    drag: false,
-    unitsView: true,
-  });
-
-  const onChange = (id: number) => {
-    dispatch({ type: "set-merge-ids", id });
-  };
-  const mergeSections = () => {
-    dispatch({ type: "merge-ids" });
-  };
-
-  const onDragEnd = (r: DropResult) => {
-    dispatch({ type: "dropped-unit", result: r });
-  };
-
-  const addUnitAt = (
-    unit: UnitEditorModel,
-    section_index: number,
-    unit_index: number
-  ) => {
-    /// callback for adding a unit in column at index i
-    // should probably have a way to split the section as well..
-    console.log("Adding At", unit, section_index, unit, unit_index);
-    dispatch({
-      type: "add-unit-at",
-      section_index,
-      unit,
-      unit_index,
-    });
-  };
-
-  const editUnitAt = (
-    unit: UnitEditorModel,
-    section_index: number,
-    unit_index: number
-  ) => {
-    /// callback for editing a unit
-    // should probably have a way to split the section as well..
-    console.log("Editing At", section_index, unit, unit_index);
-    dispatch({
-      type: "edit-unit-at",
-      section_index,
-      unit,
-      unit_index,
-    });
-  };
-
-  const headers = [
-    "",
-    "Section number",
-    "Top interval",
-    "Bottom interval",
-    "# of units",
-  ];
+  const { col_id, colSections, column, query, sections } = props;
 
   return h(BasePage, { query }, [
     h("h3", [
@@ -132,60 +66,6 @@ export default function Columns(props: {
         href: `/column/${col_id}/edit`,
       }),
     ]),
-    h(ColumnPageBtnMenu, {
-      state: {
-        unitsView: state.unitsView,
-        drag: state.drag,
-        divideIds: state.divideIds,
-        mergeIds: state.mergeIds,
-      },
-      toggleUnitsView: () => dispatch({ type: "toggle-units-view" }),
-      toggleDrag: () => {
-        dispatch({ type: "toggle-drag" });
-      },
-      divideSection: () => {},
-      mergeSections: () => {},
-    }),
-    h.if(colSections.length > 0 && !state.unitsView)(ColSectionsTable, {
-      colSections,
-      onChange,
-      headers,
-    }),
-    h.if(state.sections.length > 0 && state.unitsView)("div", [
-      //@ts-ignore
-      h(MinEditorToggle, {
-        btnText: "create new unit on top",
-        persistChanges: (e, c) =>
-          dispatch({
-            type: "add-unit-at",
-            section_index: 0,
-            unit: e,
-            unit_index: 0,
-          }),
-      }),
-      h(ColSecUnitsTable, {
-        onClickDivideCheckBox: (id: number) =>
-          dispatch({ type: "set-divide-ids", id }),
-        state,
-        onDragEnd,
-        editUnitAt,
-        addUnitAt,
-      }),
-      //@ts-ignore
-      h(MinEditorToggle, {
-        persistChanges: (e, c) =>
-          dispatch({
-            type: "add-unit-at",
-            section_index: state.sections.length - 1,
-            // an annoying way to get the index of the last unit in last section
-            unit_index: Object.values(
-              state.sections[state.sections.length - 1]
-            )[0].length,
-            unit: e,
-          }),
-
-        btnText: "create new unit on bottom",
-      }),
-    ]),
+    h(UnitSectionTable, { sections, colSections }),
   ]);
 }
