+++ dacite/src/components/unit/table/index.ts
@@ -1,0 +1,2 @@
+export * from "./table";
+export * from "./reducer";
